refactor: improved project structure
